backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.30.0
redis[hiredis]==5.2.1
httpx==0.28.1
pydantic-settings==2.7.1
beautifulsoup4==4.12.3
lxml==5.3.0
anthropic>=1.0.0
